requirements.txt
colormath==3.0.0
colour==0.1.5
Pillow==5.0.0
loguru==0.2.3